MMCM/requirements.txt
absl-py==2.1.0
accelerate==0.22.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.2.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
argparse==1.4.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
basicsr==1.4.2
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
black==23.12.1
blinker==1.7.0
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
cmake==3.28.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cos-python-sdk-v5==1.9.22
coscmd==1.8.6.30
crcmod==1.7
cryptography==41.0.7
cycler==0.12.1
cython==3.0.2
datetime==5.4
debugpy==1.8.0
decorator==4.4.2
decord==0.6.0
dill==0.3.7
docker-pycreds==0.4.0
dulwich==0.21.7
easydict==1.11
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.0
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.1
future==0.18.3
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==4.5.3
gitdb==4.0.11
gitpython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.43.2
gradio-client==0.5.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==7.0.1
importlib-resources==6.1.1
infomap==2.7.1
iniconfig==2.0.0
insightface==0.7.3
invisible-watermark==0.1.5
iopath==0.1.10
ip-adapter==0.1.0
iprogress==0.4
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.0.3
jax==0.4.23
jedi==0.19.1
jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
json-tricks==3.17.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.0
jupyter-core==5.7.1
jupyterlab-widgets==3.0.9
keras==2.12.0
kiwisolver==1.4.5
kornia==0.7.0
lazy-loader==0.3
libclang==16.0.6
librosa==0.10.1
lightning-utilities==0.10.0
lit==17.0.6
llvmlite==0.41.1
lmdb==1.4.1
loguru==0.6.0
markdown==3.5.2
markdown-it-py==3.0.0
markupsafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.3
ml-dtypes==0.3.2
mmcv==2.1.0
mmdet==3.2.0
mmengine==0.10.2
mmpose==1.3.1
model-index==0.1.11
modelcards==0.1.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.2.1
ninja==1.11.1
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
onnxsim==0.4.33
open-clip-torch==2.20.0
opencv-contrib-python==4.8.0.76
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.34
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.10
oss2==2.17.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
pillow==10.2.0
pip==23.3.1
platformdirs==4.1.0
pluggy==1.3.0
pooch==1.8.0
portalocker==2.8.2
prettytable==3.9.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.2
pydeck==0.8.1b0
pydub==0.25.1
pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
pysocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==2.0.8
pytube==15.0.0
pytz==2023.3.post1
pywavelets==1.5.0
pyyaml==6.0.1
pyzmq==25.1.2
qudida==0.0.4
redis==4.5.1
referencing==0.32.1
regex==2023.12.25
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.17.1
rsa==4.9
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==60.2.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
stack-data==0.6.3
starlette==0.35.1
streamlit==1.30.0
streamlit-drawable-canvas==0.9.3
sympy==1.12
tabulate==0.9.0
tb-nightly==2.11.0a20220906
tenacity==8.2.3
tensorboard==2.12.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminaltables==3.1.10
test-tube==0.7.5
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.9.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1+cu118
torch-tb-profiler==0.4.1
torchmetrics==1.1.1
torchvision==0.15.2+cu118
tornado==6.4
tqdm==4.65.2
traitlets==5.14.1
transformers==4.33.1
triton==2.0.0
typing-extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.18
urwid==2.4.2
uvicorn==0.26.0
validators==0.22.0
wandb==0.15.10
watchdog==3.0.0
wcwidth==0.2.13
webdataset==0.2.86
webp==0.3.0
websockets==11.0.3
werkzeug==3.0.1
wget==3.2
wheel==0.41.2
widgetsnbextension==4.0.9
wrapt==1.14.1
xformers==0.0.21
xmltodict==0.13.0
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope-interface==6.1
fire==0.6.0
xlsxwriter
git+https://github.com/tencent-ailab/IP-Adapter.git@main
git+https://github.com/openai/CLIP.git